refactor: display accepted vacancies
