+++ src/component/employer/vacancyPreview.tsx
@@ -1,15 +1,15 @@
 import Link from "next/link";
 import { format } from "date-fns";
-import type { Vacancy } from "~/utils/dbSchema/models";
+import type { Vacancy } from "~/server/db/types/schema";
 
 export function VacancyPreview({
   vacancy,
 }: {
-  vacancy: Omit<Vacancy, "employerId" | "moderationStatus">;
+  vacancy: Omit<Vacancy, "id" | "employer_id" | "moderation_status_id">;
 }) {
   return (
     <div className="w-full rounded-md border border-gray-300 p-4">
-      <Link href={`/job/${vacancy.questionnaireId}`}>
+      <Link href={`/job/${vacancy.vacancy_uuid}`}>
         <div className="mb-3 flex items-center">
           <div className="font-bold text-blue-600 hover:text-blue-800">
             {vacancy.specialty}
@@ -36,9 +36,9 @@ export function VacancyPreview({
           <strong>Умови:</strong> {vacancy.conditions}
         </div>
       )}
-      {vacancy.workSchedule && (
+      {vacancy.work_schedule && (
         <div className="mb-2">
-          <strong>Графік роботи:</strong> {vacancy.workSchedule}
+          <strong>Графік роботи:</strong> {vacancy.work_schedule}
         </div>
       )}
       {vacancy.employment && (
@@ -48,7 +48,7 @@ export function VacancyPreview({
       )}
       <div className="mb-2">
         <strong>Опубліковано:</strong>{" "}
-        {format(new Date(vacancy.dateOfPublication), "d MMMM yyyy, HH:mm")}
+        {format(new Date(vacancy.publication_date), "d MMMM yyyy, HH:mm")}
       </div>
     </div>
   );
